+++ src/Game/FrenchCards.hs
@@ -1,1 +1,3 @@
-module Game.FrenchCards where
+module Game.FrenchCards (module Cards) where
+
+import Game.FrenchCards.Card as Cards