+++ src/Game/FrenchCards/Suit.hs
@@ -1,0 +1,5 @@
+module Game.FrenchCards.Suit where
+
+
+data Suit = C | D | H | P
+  deriving (Bounded, Enum, Eq, Ord, Show, Read)